Validaties aanpassen

+++ requests/user/delete.php
@@ -10,12 +10,20 @@
             $email = $_POST["email"];
             $validation = new validation();
             if($validation->filter_length($email)){
-                  $database = new Database();
-                  $result = $database->delete("DELETE FROM User WHERE email = ?", array($email));
-                  if($result == true){
-                        print("true");
+                  if($validation->filter_characters($email)){
+                        if($validation->validate_email($email)){
+                              $database = new Database();
+                              $result = $database->delete("DELETE FROM User WHERE email = ?", array($email));
+                              if($result == true){
+                                    print("true");
+                              } else {
+                                    print("false");
+                              }
+                        } else {
+                              print(json_encode($validation->get_errors()[0]));
+                        }
                   } else {
-                        print("false");
+                        print(json_encode($validation->get_errors()[0]));
                   }
             } else {
                   print(json_encode($validation->get_errors()[0]));
